+++ src/snarky/gen/transaction.ts
@@ -14,7 +14,10 @@ import {
   Events,
   SequenceEvents,
 } from '../transaction-leaves.js';
-import { asFieldsAndAux, AsFieldsAndAux } from '../transaction-helpers.js';
+import {
+  provableFromLayout,
+  ProvableExtended,
+} from '../transaction-helpers.js';
 import * as Json from './transaction-json.js';
 import { jsLayout } from './js-layout.js';
 
@@ -23,22 +26,22 @@ export { Json };
 export * from '../transaction-leaves.js';
 
 type CustomTypes = {
-  StringWithHash: AsFieldsAndAux<
+  StringWithHash: ProvableExtended<
     {
       data: string;
       hash: Field;
     },
     Json.TypeMap['string']
   >;
-  TokenSymbol: AsFieldsAndAux<TokenSymbol, Json.TypeMap['string']>;
-  Events: AsFieldsAndAux<
+  TokenSymbol: ProvableExtended<TokenSymbol, Json.TypeMap['string']>;
+  Events: ProvableExtended<
     {
       data: Field[][];
       hash: Field;
     },
     Json.TypeMap['Field'][][]
   >;
-  SequenceEvents: AsFieldsAndAux<
+  SequenceEvents: ProvableExtended<
     {
       data: Field[][];
       hash: Field;
@@ -252,7 +255,7 @@ type ZkappCommand = {
   memo: string;
 };
 
-let ZkappCommand = asFieldsAndAux<ZkappCommand, Json.ZkappCommand>(
+let ZkappCommand = provableFromLayout<ZkappCommand, Json.ZkappCommand>(
   jsLayout.ZkappCommand as any,
   customTypes
 );
@@ -444,7 +447,7 @@ type AccountUpdate = {
   };
 };
 
-let AccountUpdate = asFieldsAndAux<AccountUpdate, Json.AccountUpdate>(
+let AccountUpdate = provableFromLayout<AccountUpdate, Json.AccountUpdate>(
   jsLayout.AccountUpdate as any,
   customTypes
 );